Migration to Identity Management -  In progress
Dashboard Management - Complete. Testing Pending

+++ LYSAdmin.Web/LYSAdmin.Domain/DashboardManagement/IDashboardManagement.cs
@@ -9,6 +9,6 @@ namespace LYSAdmin.Domain.DashboardManagement
 {
     public interface IDashboardManagement
     {
-        DonughtChart GetDonught(int OwnerID);
+        DonughtChart GetDonught(long OwnerID);
     }
 }
